requirements.txt
# TestPilot AI — Core Dependencies

# AI
anthropic>=0.40.0

# HTTP
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-json-report>=1.5.0
respx>=0.21.0          # Mock httpx calls (Siebel REST)

# SOAP (for Siebel SOAP integration)
zeep>=4.2.1

# Config
pyyaml>=6.0.1

# Dev extras
pytest-cov>=5.0.0      # Coverage reports
